refactor(git): extract Commit tree reconciliation
